// lib/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseURL: "https://bpsportal-dev3.bankofamerica.com/service" | (string & {});
};

export type VulnerabilityCmExtraBulkUpdateModel = {
  items?: Array<VulnerabilityCmExtraUpdateItem> | null;
  id?: number;
};

export type VulnerabilityCmExtraUpdateItem = {
  gisid?: string | null;
  disposition?: string | null;
  requestedPatchWindow?: string | null;
  expectedRemediationDate?: string | null;
  crq?: string | null;
  identifiedBlockers?: string | null;
};

export type VulnerabilityCmFilterRequestModel = {
  reportDate?: string | null;
  gisid?: string | null;
  applicationId?: number | null;
  cioDisplayName?: string | null;
  workstreamObservationType?: string | null;
  severityRisk?: string | null;
  scorecardSource?: string | null;
  deviceType?: string | null;
  operatingEnvironment?: string | null;
  pageNumber?: number;
  pageSize?: number;
};

export type GetApiV2VulnerabilityCmVulnerabilityFilterOptionData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/api/v2/VulnerabilityCm/VulnerabilityFilterOption";
};

export type GetApiV2VulnerabilityCmVulnerabilityFilterOptionResponses = {
  /**
   * OK
   */
  200: unknown;
};

export type PostApiV2VulnerabilityCmBulkUpdateVcmExtraData = {
  body?: VulnerabilityCmExtraBulkUpdateModel;
  path?: never;
  query?: never;
  url: "/api/v2/VulnerabilityCm/BulkUpdateVCMExtra";
};

export type PostApiV2VulnerabilityCmBulkUpdateVcmExtraResponses = {
  /**
   * OK
   */
  200: unknown;
};

export type PostApiV2VulnerabilityCmGetVulnerabilityCmData = {
  body?: VulnerabilityCmFilterRequestModel;
  path?: never;
  query?: never;
  url: "/api/v2/VulnerabilityCm/GetVulnerabilityCM";
};

export type PostApiV2VulnerabilityCmGetVulnerabilityCmResponses = {
  /**
   * OK
   */
  200: unknown;
};